requirements.txt
numpy>=1.26.4         
pandas>=2.2.1         
scikit-learn==1.6.1            
scipy>=1.13.0

xgboost>=2.1.0                 
lightgbm>=4.3.0

mlflow>=2.12.1                 
psycopg2-binary>=2.9.9

shap>=0.45.0

fastapi>=0.111.0             
uvicorn[standard]>=0.29.0      
pydantic>=2.7.1                
httpx>=0.27.0

dash>=2.17.0
plotly>=5.22.0
requests>=2.31.0

pyyaml>=6.0.1

pytest>=8.2.0
pytest-cov>=5.0.0

jupyter>=1.0.0
ipykernel>=6.29.4
matplotlib>=3.8.4
seaborn>=0.13.2
optuna>=3.6.0